Relocates CharacterConfiguration to nucypher.config.base
