+++ tests/integration/config/test_character_configuration.py
@@ -35,7 +35,7 @@ from nucypher.config.characters import (
 )
 from nucypher.config.constants import TEMPORARY_DOMAIN
 from nucypher.config.keyring import NucypherKeyring
-from nucypher.config.node import CharacterConfiguration
+from nucypher.config.base import CharacterConfiguration
 from nucypher.config.storages import ForgetfulNodeStorage
 
 # Main Cast
